requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appdirs=1.4.4=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
black=19.10b0=py_0
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.10.8=h033912b_0
certifi=2021.10.8=py39h6e9494a_1
click=8.0.3=pyhd3eb1b0_0
ffmpeg=4.3=h0a44026_0
freetype=2.11.0=hd8bbffd_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
icu=58.2=h0a44026_3
intel-openmp=2021.4.0=hecd8cb5_3538
isort=5.10.0=pyhd8ed1ab_0
jpeg=9d=h9ed2024_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
libpng=1.6.37=ha441bb4_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=h87d7836_0
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libwebp=1.2.0=hacca55c_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=hcdb78fc_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
mypy_extensions=0.4.3=py39hecd8cb5_0
ncurses=6.3=hca72f7f_0
nettle=3.7.3=h230ac6f_1
numpy=1.21.2=py39h4b4dc7a_0
numpy-base=1.21.2=py39he0bd621_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1l=h0d85af4_0
pathspec=0.7.0=py_0
pillow=8.4.0=py39h98e4679_0
pip=21.2.4=py39hecd8cb5_0
python=3.9.6=h88f2d9e_1
python_abi=3.9=2_cp39
pytorch=1.10.0=py3.9_0
readline=8.1=h9ed2024_0
regex=2021.8.3=py39h9ed2024_0
setuptools=58.0.4=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hce871da_0
tk=8.6.11=h7bc2e8c_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.10.0=py39_cpu
torchvision=0.11.1=py39_cpu
typed-ast=1.4.3=py39h9ed2024_1
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
